+++ realtime/viewer_manager.go
